requirements.txt
numpy
serial
opencv-python
kociemba
tk
Pillow

